chore: Renames section item stylesheet file
